detection/requirements.txt
numpy
opencv-python
imutils
Pillow
